+++ R/run_sim.R
@@ -13,7 +13,8 @@ run_sim <- function(model_config) {
   # Time unit  = minutes
   env <- simmer("pathway")
 
-  # create the distribution functions
+  #### create the distribution functions ####
+  ## continuous distributions ##
   # patient arrivals
   rate <- mc$pat_referral_rate * 12 / 365 / 24 / 60 # monthly -> annual, then calculate patients per minute
   dist_patient_arrival <- function() rexp(1, rate)
@@ -31,6 +32,11 @@ run_sim <- function(model_config) {
 
   dist_post_op_ward_los <- function() rexp(1, 1 / (60 * 24 * mc$post_op_los)) # 60x24 = days
   # dist_post_op_ward_los()
+
+  ## discrete distributions ##
+  # OP did not attend (DNA) rate. 0 = attended, 1 = did not attend
+  dist_op_dna <- function() sample(0:1, 1, FALSE, c(100 - mc$op_dna_rate, mc$op_dna_rate))
+  dist_op_dna()
 
   # OP outcoming result. 1 = admit to wl, 2 = OP followup, 3 = discharged
   op_disch_rate <- (100 - mc$op_admit_rate - mc$op_fup_rate)
@@ -51,35 +57,46 @@ run_sim <- function(model_config) {
   )
 
   # create the patient pathway branches
+  branch_op_dna <- trajectory("op did not attend") |>
+    # the dna consumes the same clinic resource as an attendance
+    log_("OP: DNA") |>
+    set_attribute("OP appt DNA", 1) |>
+    timeout(mc$op_clinic_length) |>
+    release("OP Clinic", 1) |>
+    rollback("op_clinic") # rollback to tagged resource
+
   branch_discharge_from_op <- trajectory("discharged from OP appt") |>
-    set_attribute("discharged home", 1) |>
-    log_("Discharged from OP appt")
+    log_("OP outcome: Discharge") |>
+    set_attribute("discharged home", 1)
 
   branch_followup_later <- trajectory("book OP followup") |>
+    log_("OP outcome: Follow-up later") |>
     set_attribute("OP_fup_booked", 1) |>
-    log_("Follow-up OP appt booked") |>
     rollback("op_clinic") # rollback to tagged resource
 
   branch_admit <- trajectory("admit for treatment") |>
+    log_("OP outcome: Admit") |>
     set_attribute("admitted_for_treatment", 1) |>
     # take a pre-op bed
-    set_attribute("moved_to_pre_op_bed", 1) |>
     seize("Bed") |>
+    log_("Pre-op bed") |>
+    set_attribute("moved_to_pre_op_bed", 1) |>
     timeout(dist_pre_op_ward_los) |>
     release("Bed") |>
     # operate
-    set_attribute("moved_to_theatre", 1) |>
     seize("Theatre") |>
+    log_("Theatre") |>
+    set_attribute("moved_to_theatre", 1) |>
     timeout(dist_operating_time) |>
     release("Theatre") |>
-    log_("Im recovering") |>
     # take a recovery ward bed
-    set_attribute("moved_to_post_op_bed", 1) |>
     seize("Bed") |>
+    log_("Post-op bed") |>
+    set_attribute("moved_to_post_op_bed", 1) |>
     timeout(dist_post_op_ward_los) |>
     release("Bed") |>
-    set_attribute("discharged home", 1) |>
-    log_("Discharged from bed")
+    log_("IP discharged") |>
+    set_attribute("discharged home", 1)
 
 
   # create the overall patient pathway
@@ -87,11 +104,24 @@ run_sim <- function(model_config) {
     #  log_("Referred in") |>
     ## add an intake activity
     seize("OP Clinic", 1, tag = "op_clinic") |>
-    timeout(function() rnorm(1, mean = mc$op_clinic_length, sd = 6)) |>
+
+    # create a branch to model OP DNAs
+    branch(
+      dist_op_dna,
+      continue = FALSE,
+      branch_op_dna
+    ) |>
+
+    # if no DNA, continue with the OP appointment
+    log_("OP: Attended") |>
+    set_attribute("OP appt attended", 1) |>
+    timeout(mc$op_clinic_length) |>
     release("OP Clinic", 1) |>
+
     # branch into admission and discharge
     branch(
-      dist_op_outcome, FALSE,
+      dist_op_outcome,
+      continue = FALSE,
       branch_admit,
       branch_followup_later,
       branch_discharge_from_op